Moved profilecontroller code to usercontroller, removed reduntant endpoints, added an id parameter for updating profile, and made backend/frontend interface code compatible
