+++ backend/src/controllers/UserController.ts
@@ -1,14 +1,74 @@
 import crypto from "crypto";
-import { Document, Model, ObjectId } from "mongoose";
+import { Document, Model, ObjectId, ClientSession } from "mongoose";
 import { IUser } from "../models/User";
 import { GithubProfile, parseAvatarUrl } from "../auth/github";
 import NotFoundError from "./errors/NotFoundError";
+import UnauthorizedError from "./errors/UnauthorizedError";
+import UnexpectedError from "./errors/UnexpectedError";
 
 export type UserRecord = IUser & Document<ObjectId, any, IUser>;
 
-class UserController {
-  constructor(private userModel: Model<IUser>) {}
+export interface ProfileUpdateParams {
+  username?: string;
+  displayName?: string;
+  techs?: string[];
+}
 
+class UserController {
+  constructor(
+    private userModel: Model<IUser>,
+    private createSession: () => Promise<ClientSession>
+  ) {}
+  //TODO: Update User
+  async updateProfile(
+    id: string,
+    paramsId: string,
+    params: ProfileUpdateParams,
+    isAdmin?: boolean
+  ) {
+    await this.validateUserUpdatePermission(id, isAdmin);
+    const session = await this.createSession();
+    let user: null | Document<IUser> = null;
+    try {
+      user = await this.userModel.findOneAndUpdate(
+        {
+          _id: id,
+        },
+        {
+          _id: id,
+          ...params,
+        },
+        {
+          new: true,
+          upsert: true,
+          session,
+        }
+      );
+    } finally {
+      session.endSession();
+    }
+
+    if (null !== user) {
+      //await (user as Document<IUser>).populate("user");
+      await (user as Document<IUser>).populate("techs");
+      return user;
+    } else {
+      throw new UnexpectedError("Expected user to be updated");
+    }
+  }
+  private async validateUserUpdatePermission(id: string, isAdmin?: boolean) {
+    isAdmin = isAdmin || false;
+
+    if (!isAdmin) {
+      const user = await this.userModel.findOne({
+        _id: id,
+      });
+
+      if (!user || user.id !== id) {
+        throw new UnauthorizedError("Updater is not the user");
+      }
+    }
+  }
   async createOrUpdateUser(githubId: number): Promise<UserRecord> {
     return await this.userModel.findOneAndUpdate(
       { githubId },
