+++ frontend/src/components/user_page/UserSettingsPage.tsx
@@ -4,43 +4,21 @@ import UserSettingsLayout from "./UserSettingsLayout";
 //import UpdateUserForm from "./UpdateUserForm";
 
 const UserSettingsPage: FunctionComponent = () => {
-  // const [isLoading, setIsLoading] = useState(true);
-  const [user, setUser] = useState<any>();
   const [loading, setLoading] = useState(false);
 
   useEffect(() => {
-    let isMounted = true;
-    setLoading(true);
+    //let isMounted = true;
+    setLoading(false);
 
-    fetch("/api/v1/users").then(async (response) => {
-      if (response.status === 200) {
-        if (isMounted) {
-          const data = await response.json();
-
-          setUser(data);
-          setLoading(false);
-        }
-      } else {
-        console.error(
-          "failed to load user",
-          response.status,
-          await response.json()
-        );
-      }
-    });
     return () => {
-      isMounted = false;
+      //isMounted = false;
     };
   }, []);
 
   return (
     <>
       <div className="w-full bg-medGray">
-        {!loading ? (
-          <UserSettingsLayout userProps={user} />
-        ) : (
-          <LoadingSpinner />
-        )}
+        {!loading ? <UserSettingsLayout /> : <LoadingSpinner />}
       </div>
     </>
   );
